requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
Pillow==10.0.0
Werkzeug==2.3.6
python-dotenv==1.0.0
Flask-Session==0.5.0
WTForms==3.0.1
email-validator==2.0.0
gunicorn==21.2.0